requirements.txt
cryptography==40.0.1
pycodestyle==2.10.0
pycryptodomex==3.17
PyJWT==2.6.0
PyYAML==6.0
